rltool/state: fix generalizeGoals with cterm patterns
Previously, when generalizeGoals used a constrained term pattern of
the form `(T | C)`, the pattern was transformed into `(T | C /\ F)`
where F is a fresh formula variable and where `/\` is ACU with identity
`true`. If the extension variable `F` matched a formula term φ, then
we cannot blindly do the following: `(T | C /\ F = φ)` because φ
is not a term in our object theory, which was enough for
generalizeGoals in the case of patterns of the form `T`.

To fix the bug, we handle the above case by applying the inverse
functor `eqform-func-op` so that we obtain `(T | C /\ φ)`.