rltool/ibos: unconditionalize SOP predicates
This was an ill-fated attempt to speed things up, but it didn't work
out. This commit is being added for documentation purposes and will be
reverted later. Don't do this, because unconditionalizing predicates
causes them to blow up into huge terms during evaluation which either
require specialized reduction rules to handle or blow up into a huge
number of clauses.