rltool/conrew: prove clauses valid by contradiction
If a clause (∧ᵢ uᵢ = vᵢ) => (∨ₖ sₖ = tₖ) is such that the antecedent
already contains a provable contradiction (i.e. that for two, ground
constructor terms α and β that are known α != β, we can prove that
α = β), then we prove the clause is valid.

TODO: Currently, the tool hardcodes α and β to be the Boolean terms true
and false. This should be properly parameterized in a future version.