rltool/{congclos,conrew}: fix subtle ordering bug
Add a boolean flag to the congruence closure allowing to choose a fast
term ordering algorithm that does not necessarily agree with the
termination order of the module that it is based on or a slower
algorithm that does agree.

Previously, the congruence closure algorithm used an optimization in the
term ordering based on term size. However, when congruence closure is
used as part of contextual rewriting, it is important that our ordering
agrees with the termination ordering for the module.