rltool/meta: add fixed sig identifier functor
Occasionally, it is useful to have an unbounded number of fresh
identifiers available but in a fixed signature. We accomplish
this by adding a fresh copy of the natural numbers to our signature and
then fresh injections from that copy into each sort.

We also provide mapping functions to convert variables into these fresh
identifers and vice versa.