rltool/ibos: add more complex version of proc-loaded-ind?
Note this commit exists for documentation reasons and will be reverted.
The strategy of enriching proc-loaded-ind? is too complex and doesn't
ensure a one-to-one mapping between a NetProc and WebProc that is
loaded so it still can't force the needed correspondence between the
`proc-loaded?` and `proc-loaded-ind?` predicates.

The revised strategy is to enforce no duplication on process metadata on
the inner URL position, which will force the WebProc <-> NetProc
metadata mapping to be bijective.