FIXME update to proof script/pred spec
NOTE: false case for SOP pred underspecified --- need to search for
label everywhere and return false if we find it...

Or do we? The single process case didn't need this rule.

Maybe it's just a matter of completing the process group and then
things can be thrown out... let's see how that goes first